requirements.txt
Django>=4.1,<4.3
djangorestframework>=3.11,<3.15
drf-spectacular>=0.27,<0.28
Pillow>=10.1,<11.1
psycopg2>=2.9.3,<2.10
djangorestframework-simplejwt>=5.3,<5.4
requests==2.32.2
django-crispy-forms<=2.4
django-ckeditor==6.7.2
crispy-bootstrap5<=2025.4
django-formset==1.6.1